Unify progress step rows and remove UsedToolsList
- Replace separate completedToolRow/runningToolRow/blockedToolRow with
  unified StepDetailRow component that handles all states
- Add expandable detail section with technical info, screenshots, and
  output for completed tools
- Prefer reasonDescription over friendlyRunningLabel for running tools
- Delete UsedToolsList.swift (functionality absorbed into StepDetailRow)
